refactor: extract selfCloseVoidElements into HtmlSanitizer lib
Move the HTML5 void-element sanitizer out of ChapterHtmlSlimParser
into a standalone lib/HtmlSanitizer utility. The scanner is a generic
HTML-to-XML pre-pass, not chapter-parser logic, and its presence
bloated an already dense parser file.

Pure move: state machine is byte-identical (only the log tag changed
from EHP to HSAN). Caller in Section.cpp rewired; Memory.h include
dropped from the parser as no longer used.
